refactor livecd-install.sh and prompt all questions at beginning

+++ bin/livecd-install.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 
-# livecd-install.sh v0.1
+# livecd-install.sh v0.2
 #
 # Copyright (C) 2023 Kristofer Berggren
 # All rights reserved.
@@ -28,24 +28,39 @@ warn() {
   >&2 echo "${1}"
 }
 
-user_check() {
-  [[ "$(whoami)" == "root" ]] || exiterr "This script must be run as root"
-  echo "WARNING: This script will erase all content on first hard drive."
-  echo "It will not provide options to select which drive or partition to use."
-  read -p "Proceed to install (yes/no)? "
-  [[ "${REPLY}" == "yes" ]] || exiterr "User abort"
-}
-
 DEBUG=$([[ "${1}" == "-d" ]] && echo "1" || echo "0")
 MNTDISK="/mnt/disk"
 MNTROOT="/mnt/root"
 ROOT="/"
 USER="user"
+PASS="pass"
 
-install_void() {
-  user_check
-  [[ -x "$(command -v parted)" ]] || xbps-install -y parted || exiterr "Failed to install parted"
-  [[ -x "$(command -v rsync)" ]] || xbps-install -y rsync || exiterr "Failed to install rsync"
+user_check() {
+  [[ "$(whoami)" == "root" ]] || exiterr "This script must be run as root"
+  echo "WARNING: This script will erase all content on first hard drive."
+  echo "It will not provide options to select which drive or partition to use."
+  read -p "Proceed to install (y/n)? "
+  [[ "${REPLY}" == "y" ]] || exiterr "User abort"
+  read -p "Create new user (${USER}): "
+  [[ "${REPLY}" != "" ]] && USER="${REPLY}"
+  read -s -p "Set user password (${PASS}): "
+  [[ "${REPLY}" != "" ]] && PASS="${REPLY}"
+}
+
+distro_check() {
+  if [[ "${DISTRO}" != "Void" ]]; then
+    exiterr "unsupported linux distro (${NAME})"
+  fi
+}
+
+setup_deps() {
+  if [[ "${DISTRO}" == "Void" ]]; then
+    [[ -x "$(command -v parted)" ]] || xbps-install -y parted || exiterr "Failed to install parted"
+    [[ -x "$(command -v rsync)" ]] || xbps-install -y rsync || exiterr "Failed to install rsync"
+  fi
+}
+
+cleanup_mounts() {
   [[ -d "${MNTDISK}/dev" ]] && (umount ${MNTDISK}/dev || warn "Unmount /dev failed")
   [[ -d "${MNTDISK}/proc" ]] && (umount ${MNTDISK}/proc || warn "Unmount /proc failed")
   [[ -d "${MNTDISK}/sys" ]] && (umount ${MNTDISK}/sys || warn "Unmount /sys failed")
@@ -53,6 +68,13 @@ install_void() {
   [[ -d "${MNTDISK}" ]] && (rmdir ${MNTDISK} || exiterr "Delete mount point ${MNTDISK} failed")
   [[ -d "${MNTROOT}" ]] && (umount ${MNTROOT} || exiterr "Unmount ${MNTROOT} failed")
   [[ -d "${MNTROOT}" ]] && (rmdir ${MNTROOT} || exiterr "Delete mount point ${MNTROOT} failed")
+}
+
+perform_install() {
+  # Clean up existing mounts (to support retry running script)
+  cleanup_mounts
+
+  # Set up hard drive
   DRIVEINFO=$(lsblk -io TYPE,KNAME,SIZE | grep "^disk" | sort -k3,3hr | head -1)
   [[ "${DRIVEINFO}" != "" ]] || exiterr "No drive found"
   DEVICE="/dev/$(echo "${DRIVEINFO}" | awk '{print $2}')"
@@ -64,44 +86,39 @@ install_void() {
   (mkdir ${MNTDISK} && mount ${PARTITION} ${MNTDISK}) || exiterr "Mount ${PARTITION} failed"
   (mkdir ${MNTROOT} && mount ${ROOT} ${MNTROOT} -o bind) || exiterr "Mount ${ROOT} failed"
 
+  # Copy Live-image to hard drive
   rsync -a --info=progress2 --no-i-r ${MNTROOT}/ ${MNTDISK}/
-  # rsync < 3.1 may need: rsync -aP ${MNTROOT}/ ${MNTDISK}/
 
+  # Install and update Grub
   unset UUID
   eval $(blkid | grep "^${PARTITION}" | awk '{print $2}')
   echo "UUID=${UUID}	/	ext4	defaults	0	2" >> ${MNTDISK}/etc/fstab
   grub-install --root-directory=${MNTDISK} ${DEVICE}
-
   mount --bind /dev ${MNTDISK}/dev || exiterr "Mount /dev failed"
   mount --bind /proc ${MNTDISK}/proc || exiterr "Mount /proc failed"
   mount --bind /sys ${MNTDISK}/sys || exiterr "Mount /sys failed"
-
   chroot ${MNTDISK} update-grub || exiterr "Update Grub failed"
 
+  # Add user
   chroot ${MNTDISK} useradd --create-home ${USER} || exiterr "User add failed"
   chroot ${MNTDISK} usermod --append --groups wheel ${USER} || exiterr "User mod failed"
-  echo "Set password for '${USER}':"
-  chroot ${MNTDISK} passwd ${USER} || exiterr "User password set failed"
+  yes "${PASS}" | chroot ${MNTDISK} passwd ${USER} || exiterr "User password set failed"
   chroot ${MNTDISK} sh -c "echo \"%wheel ALL=(ALL:ALL) NOPASSWD: ALL\" >> /etc/sudoers" || exiterr "Add to sudoers failed"
 
+  # Cleanup
   if [[ "${DEBUG}" == "0" ]]; then
-    [[ -d "${MNTDISK}/dev" ]] && (umount ${MNTDISK}/dev || exiterr "Unmount /dev failed")
-    [[ -d "${MNTDISK}/proc" ]] && (umount ${MNTDISK}/proc || exiterr "Unmount /proc failed")
-    [[ -d "${MNTDISK}/sys" ]] && (umount ${MNTDISK}/sys || exiterr "Unmount /sys failed")
-    [[ -d "${MNTDISK}" ]] && (umount ${MNTDISK} || exiterr "Unmount ${MNTDISK} failed")
-    [[ -d "${MNTDISK}" ]] && (rmdir ${MNTDISK} || exiterr "Delete mount point ${MNTDISK} failed")
-    [[ -d "${MNTROOT}" ]] && (umount ${MNTROOT} || exiterr "Unmount ${MNTROOT} failed")
-    [[ -d "${MNTROOT}" ]] && (rmdir ${MNTROOT} || exiterr "Delete mount point ${MNTROOT} failed")
+    cleanup_mounts
   fi
 }
 
 unset NAME
 eval $(grep "^NAME=" /etc/os-release 2> /dev/null)
-if [[ "${NAME}" == "Void" ]]; then
-  install_void
-else
-  exiterr "unsupported linux distro (${NAME})"
-fi
+DISTRO="${NAME}"
+
+distro_check
+user_check
+setup_deps
+perform_install
 
 echo ""
 echo "Completed successfully."
